smarteco/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
numpy==2.1.3
pandas==2.1.0
scikit-learn==1.3.2
joblib==1.3.2
skl2onnx==1.16.0
onnxruntime==1.16.3
pymongo==4.6.1

